Fix post merge issues
